+++ src/NexusMods.App.UI/NexusMods.App.UI.csproj
@@ -780,13 +780,13 @@
         <Compile Update="Pages\CollectionDownload\Dialogs\PremiumDownloads\DialogPremiumCollectionDownloadsView.axaml.cs">
            <DependentUpon>DialogCollectionPublishedView.axaml</DependentUpon>
         </Compile>
-        <Compile Update="Pages\DownloadsPage\DownloadsPageViewModel.cs">
+        <Compile Update="Pages\Downloads\DownloadsPageViewModel.cs">
            <DependentUpon>IDownloadsPageViewModel.cs</DependentUpon>
         </Compile>
-        <Compile Update="Pages\DownloadsPage\DownloadsPageDesignViewModel.cs">
+        <Compile Update="Pages\Downloads\DownloadsPageDesignViewModel.cs">
            <DependentUpon>IDownloadsPageViewModel.cs</DependentUpon>
         </Compile>
-        <Compile Update="Pages\DownloadsPage\DownloadsPageView.axaml.cs">
+        <Compile Update="Pages\Downloads\DownloadsPageView.axaml.cs">
            <DependentUpon>DownloadsPageView.axaml</DependentUpon>
            <SubType>Code</SubType>
         </Compile>
